[dev-mdast-editor-example] Refactor: merge node classes into their extension modules
The nodes and the extensions that configure them always travel
together, so CollapsibleNode.ts, KbdNode.ts, and FootnoteNodes.ts fold
into MdastCollapsibleExtension.ts, MdastKbdExtension.ts, and
MdastFootnoteExtension.ts respectively — the extension carries the
configuration, so its module name survives. No behavior change;
imports updated and the footnote module's re-export block became
direct exports.
